Suppress all clippy lints in generated code
Summary: When writing a code generator, it tends to be more relevant that the code generator is easily understandable and maintainable, than that the generated code is maximally idiomatic. In the clippy warnings shown in the test plan, yes we could make the Thrift templates more complex to sidestep those lints, but that wouldn't be a good tradeoff.

Reviewed By: zertosh

Differential Revision: D40314859

fbshipit-source-id: 45e1a2b2e00689942ce7bd677a7b9a87f8f21a74