Ensure field information is always available for debuggers
Summary:
Follow up to D39328395 (https://github.com/facebook/fbthrift/commit/a714a73e62718a9d5fc4ccce11fc1e832a45e41c).

We want the data in these fields to be accessible to debuggers at runtime, but
currently the unused members are getting stripped by the linker.

The compiler puts each static data field into its own section and the linker is
then able to remove sections if they are entirely unused. By forcing all of the
static data members into the same section, the unused ones are not able to be
stripped by the linker as they are not in an unused section.

Reviewed By: iahs

Differential Revision: D40232436

fbshipit-source-id: 431c28e7a71b815cb82b7505829597f4d3ee86dc