Enable list and map element patching
Summary:
This diff addresses containers of contaners codegen issue for the
patch. New types will be generated for containers that are values or another
container. E.g. if there is a field `6: list<map<string, int>> foo` in struct
called MyStruct. The patch for the list will have name MyStructField6Patch. And
corresponding patch struct for the map inside that list will have type
MyStructField6PatchElementPatch. If there are multiple level of nestedness
ElementPatch suffix is added for every level.

Reviewed By: Alfus

Differential Revision: D40529467

fbshipit-source-id: 952576b9569e72414b1a5e777502a0e93b9a4cde