Report cyclic dependencies in thrift compiler
Summary:
Proactively fail C++ generation if the generated code is likely to fail to compile with declaration order issues. These issues produce cryptic error messages that are difficult to debug, so failing early will be a substantial devX win.
The caveat is that there are some niche cases where the thrift compiler thinks some code won't compile but it will (e.g. if an adapter copies all the fields of a struct into a pure-C++ struct instead of storing the thrift struct). These are already usually broken by the topo sort, but there is a risk some previously-working files will break now. We should address these by adding information to the thrift compiler's view instead of the current behavior that depends on the declaration order in the .thrift file.

Reviewed By: Alfus

Differential Revision: D40521607

fbshipit-source-id: d2a7c90e6b7f91422a89136a0fd7a86dcd4ec1a1