@thrift.InternBox
Summary:
thrift.InternBox for unqualified (Fill) field. Since this is unqualified (Fill) field, we can assume that the field can never be null.
It currently only supports Thrift struct field. When a parent struct is constructed, the intern boxed field points to the static leaky singleton that is generated from `op::getDefault`. When user clears the struct, it will points back shared default value.

It provides same Fill semantic with the current unqualified field that can distinguish absent or empty.

Reviewed By: Alfus

Differential Revision: D40423477

fbshipit-source-id: 645f2b00c3eb6ee9f077c3d9364e369a156a1546