turn off -r for thrift-python build fixtures
Summary:
The thrift compiler is obviously buggy in recursive mode, and in the case of thrift-python, constant struct nodes may be mistakenly identified as map node and the behavior is also not consistent across different runs. Since we never use recursive compile in prod, it makes no sense to turn it on for build_fixtures. The other languages may still want it so here I am turning it off for thrift-python.

Not the remove fixtures are all from annotations, which probably shouldn't been generated in fixtures in the first place.

Reviewed By: ParalaxRus

Differential Revision: D40578215

fbshipit-source-id: 02c92985fed6d75147fdf5ec94d3616f5ae77783