Log new peeking_manager.tls connection event
Summary:
When clients use TLS without providing an appropriate ALPN, `ThriftServer` uses `TransportPeekingManager` to decorate the transport and peek at the first bytes of the connection to decide on which Thrift transport to use to handle it.

We'd like to log those events so we can fix clients to provide an ALPN.

We hope to eventually be able to remove this peeking logic and enforce the use of ALPN for TLS connections.

Reviewed By: AjanthanAsogamoorthy

Differential Revision: D40524390

fbshipit-source-id: b0cc791b63d03a794e34b5812a79933ba66f7d68