Prevent generated types from generating shapes
Summary:
The `shapes` option is causing issues in www. The TL;DR here is that apply patches to legacy schemas that use the `shapes` option forces us to enable `shapes` on `patch.thrift` as well. This eventually works its way down to `standard.thrift`, which contains recursive structs that _cannot_ use the `shapes` option, because shapes cannot be recursive in Hack.

This diff just universally disables generation of `__toShape` and `__fromShape` for generated structs. These are new, and the number of conflicts in www is basically zero right now.

Reviewed By: joshkehn

Differential Revision: D41100843

fbshipit-source-id: f26b6ecdabe3c89371e5d937c8a55b934126e7d8