Do not resolve typedefs in scope_validator
Summary:
Typedefs are distinct types, with different URIs than the underlying types, which means using them as structured annotations will not register as use of the underlying annotation. However the scope validator resolves typedefs, so no warning will be produced about this. Stop doing that so warnings (and later errors) will be raised.
Note that placeholder_typedefs aren't a concern here because a separate validator ensures that consts (which include structured annotations) have concrete types at the point of definition.

Reviewed By: vitaut

Differential Revision: D41500001

fbshipit-source-id: f0fe2f5ab9036f02b4e9c8cae7888b4a082723c7