thrift/rust: try to resolve literal string quoting problem
Summary:
Reported: https://fb.workplace.com/groups/rust.language/posts/9985872874794571/

Cope with the fact that string literals are in the form they appear in the Thrift source,
with uninterpreted `\` quote sequences. These are expected to be C-like enough that
`rustc` will accept them as-is.

We still quote things like embedded newlines and tabs rather than having to deal with
multiline strings.

#forcetdhashing

Reviewed By: iahs

Differential Revision: D41193185

fbshipit-source-id: 9c9f480933575fb556443c48d90d461cf08dfa18