Remove problematic time types
Summary:
`types.Time` is an alias of `standard.TimeStruct` but without proper adapters that map it to a native representation of a duration type it doesn't make much sense. In fact, it is harmful because users might start relying on it being represented as a struct making it hard to migrate in the future. In general we should avoid adding such half-baked functionality in the Thrift library. Same for `Timestamp`.

`minTimestamp` and `maxTimestamp` are even worse because they are not even marked as experimental.

Remove all of these until someone decides to implement this functionality properly.

Reviewed By: Mizuchi

Differential Revision: D41594879

fbshipit-source-id: 662f82571aefd4d980011eab7a3e1d66cffc26d6