Switch to core type definitions on ThriftStructTypes
Summary: We are moving non-abstract type constants out of `IThriftStruct` to a separate utility class that is not extended by all Thrift struct implementations. Currently HHVM does not adequately optimize the type structures for these definitions by de-duping them within the same unit, which is problematic when working with potentially recursive structures.

Reviewed By: prasad223

Differential Revision: D41708557

fbshipit-source-id: 841db7d8f0ac1ff09f1af491d85e49571d4eaccf