Backout @thrift.TerseWrite for FieldPatch
Summary: Terse field is required to be cleared before the deserialization due to the semantic correction, which is expensive for a large nested patch struct. thrift.InternBox has cheap clear, as it simply reseting to the shared default value (especially if the intern box is in ConstUnowned state). Backing this out temporarily before addressing build issue for thrift.InternBox.

Differential Revision: D41893706

fbshipit-source-id: b7f0fd584b203dd280d13b33fd75912bf1c85ea2