Cleanup standard.thrift
Summary: This is a continuation of D41594879 (https://github.com/facebook/fbthrift/commit/14c45d8933c43f66b2d3b095d76a21dc5800bedf). `standard.thrift` contains a number of types that are not used anywhere except tests. These types don't make much sense in their current form since most of them are not strong types and are not properly adapted (see TODOs). In fact, they are harmful because users might start relying on their current representation making it hard to migrate them to something more useful in the future (this is already happened to Duration in cws which was fixed elsewhere). In general we should avoid adding such half-baked functionality in the Thrift library. Remove these types for now.

Reviewed By: iahs

Differential Revision: D41881596

fbshipit-source-id: 4f4a804ebe379afc9e215c18dd906949f2eb424b