Add pluggable func to set custom sockopts for every client transport
Summary:
We want better tracking in fbflow for l7 protocols, the easiest way to do this is just mark the packet in userspace and have bpf read the sockopt.

Plumbing this through in thrifts client channel. Need to dbl check this actually works (not sure all clients pass in connected transports)

Reviewed By: frqiu

Differential Revision: D41185506

fbshipit-source-id: 6016c69046bd51d32a880f8f150481d966436ca5