Traverse typedefs when searching for adapter annotation
Summary:
This fixes a bug in Thrift-python compiler that ignores adapters on the intermediate typedefs.

In particular, when we have following typedefs

```
python.Adapter{
  ...
}
typedef i32 AdaptedInt
typedef AdaptedInt OtherAdaptedInt
```

The compiler would ignore the adapter annotation when using `OtherAdaptedInt` as it only looks up the annotation on the outer typedef (`OtherAdaptedInt`) and then uses the resolved type directly.

With this diff, we start traversing the typedefs to find first adapter annotation.

Reviewed By: junzh0u

Differential Revision: D42297938

fbshipit-source-id: 2f29e44bdb00d31373869e9002c77d7172410392