Fix terse struct serialization optimization by bookkeeping lastWrittenField
Summary:
`lastWrittenField_` should be tracked similar to `lastFieldId_` in order to deal with deeply nested (>=3) level of terse structs. Add it to the `lastField_` stack and track with `lastFieldId_`.

For context, `lastWrittenField_`  is required due to terse struct serialization optimization where we do passthrough emptycheck and rewind after serialization. Previously, we only call `writeFieldBegin` if the empty condition is met, but we need to keep track field id that was actually written with `lastWrittenField_`.

Differential Revision: D42563005

fbshipit-source-id: 5dadae9e389cfc37e6246f8a0064e4de992ef09f