defer all cpp2 union setter compilations
Summary:
Closes: https://github.com/facebook/fbthrift/issues/508.

As reported by https://github.com/alex-700, the field-setter codegen for unions with some indirectly-recursive fields is technically incorrect.

From https://github.com/facebook/fbthrift/issues/508#issue-1292304221:

> The copy-ctor of vector<ThriftConstValuePair> inside set_cvmap can't be instantiated before definition of ThriftConstValuePair. And a compiler, IMHO, can make a decision to instantiate it eagerly in instantiation point at the end of ThriftConstValue definition. So, IIYC, the generated code is "ill-formed no diagnostic required".

This affects fbthrift's own metadata types on recent platforms.

> The code generated for metadata.thrift is not compiling in "clang-14 & libstdc++-12.1 & c++20" environment. The version of clang isn't important, but version of libstdc++ and cxx_standard are.

The reporter offers two suggested fixes, the first of which would work well and the second of which would, owing to Hyrum's Law, possibly not work well in practice.

Here, we take a third approach, which is to ensure that every setter is a function template. We already take this approach for the field-ref accessors, so it seems natural to extend the approach to the union setters. In the function template, we ensure that the types used in the field constructions are dependent. This should prevent the compiler from attempting to use the union class layout recursively before the union class body is closed.

Reviewed By: thedavekwon

Differential Revision: D42612325

fbshipit-source-id: 1a1b3a5ec7b2dea994d742492b1dcdf660279562