Make default thread name for resource pools non-empty
Summary:
The default thread name for `ThreadManager` was `ThriftSrv-pri{X}-{Y}`, which is self-documenting. With resource pools, the default prefix is empty, in which case the thread name is `.{X}{Y}` (for normal priority, `.N{Y}`), which is non-informative and non-greppable.

This diff makes the default prefix `ThriftSrv` for resource pools as well.

Reviewed By: maalbash

Differential Revision: D42840853

fbshipit-source-id: 8802b5e0d943c5a1ba7fafdd0916e447b3638cac