Fix ExecutionScope use after move issue
Summary:
User encounter use-after-move lint warning on D42243397 (https://github.com/facebook/fbthrift/commit/f15352afa27e510231070acfdeaf4f21422e946f). Fixing the issue in that diff was out of scope, since the warning was for lines that were not modified as a part of the diff.

After looking at the code, I believe this to be a legitimate issue, so I'm fixing it.

Reviewed By: maalbash

Differential Revision: D42819227

fbshipit-source-id: a4b66712a1126a87783b51b9b5eafbda3366d1fd