python: any registry (3/n) primitive type serialization (int & float)
Summary:
Introducing an API to serialize and deserialize primitive types, see thrift/lib/python/any/serializer.pyi for the detailed APIs.

This diff only added int and float support, as a proof of concept and demo of API design.
More type support coming in later diff.

A few design choices:
1. Putting them under `thrift.python.any.serializer` instead of  `thrift.python.serializer` because they're mainly for use of Thrift Any, and we don't want to encourage normal user directly using it. Container support will be added here too for same reason. If later we decide to support direct usage, we can always merge them back to `thrift.python.serializer`.
2. Both APIs can infer the type from the passed object. They also take an optional `type.Type` to fine tune types otherwise indistinguishable in Python, for example `i16` vs `i32`, etc. We don't want to expose the internal `TypeInfo`, `Type` is part of thrift standard and a reasonable choice for this API.
NOTE: in this diff, we don't check compatibility between the `TypeName` and the actual Python value/class. That could be improved in a future diff.

Reviewed By: nanshu

Differential Revision: D43284440

fbshipit-source-id: a4f9d9c7533a76ff2efd5590ddfde18ed102b5f3