Merge overlapping IDL docs
Summary:
There is a huge overlap in content between the top-level Interface Definition Language page and the IDL Reference and despite one positioning itself as a guide they have similar presentation styles. Since the latter is much better written and has fewer inaccuracies keep it and merge the missing content from the other page into it. This way we'll have one source of truth for the Thrift language.

To keep meaningful commit history, the move and link update will be done in a follow-up diff.

Reviewed By: avalonalex

Differential Revision: D43646332

fbshipit-source-id: 6647f0b745631dfe7e84f3356b773927af7777aa