add op function to compare two thrift structs recursively
Summary:
This diff added `recursive_less_than`, which won't rely on `operator<` for thrift struct at all. Instead, it always uses `RecursiveLessThan`, recursively, to check whether one thrift object is less than the other one.

This provides an easy way to implement `operator<` for any thrift struct regardless whether `operator<` is defined for sub-struct. It will be used in the next diff in prod.

# Motivation

The existing `op::less_than` won't work in the following cases:

```
struct Foo {
  1: set<i32> (cpp.template = "std::unordered_set") field;
}

struct Bar {
  1: Foo foo;
}
```

This is because

1. `Foo::operator<` won't exist since `std::unordered_set::operator<` doesn't exist.
2. `StructLessThan<Bar>` won't work since it defers to `LessThan<Foo>` (rather than recursively deferring to StructLessThan<Foo>`), which relies on `Foo::operator<`, which doesn't exist.

However this is intended to be a temporary solution. The performance of `recursive_less_than` might be worse than native `operator<`. This is way the code is under `common/thrift/cpp` instead of core thrift library. Eventually it'll be removed.

Eventually we will make all thrift structs have `operator<` (even for structs that contains `std::unordered_set`, in which case we move elements in a `std::vector` and sort them manually). But that requires codemod and it can't be done in short term

Reviewed By: iahs

Differential Revision: D43569615

fbshipit-source-id: 8fbdc0466ad35aa3feb55db03cb30bfcbc4f5aa5