make TypeInfo.get take in a `const TypeInfo&`
Summary: So the `.get` function may access the current typeInfo for custom logic based on the current type. This may not be useful for use cases where TypeInfo for each struct is statically generated (e.g. in C++), but it would be useful for cases where dynamic TypeInfo/StructInfo is generated at runtime (e.g. Python).

Reviewed By: vitaut

Differential Revision: D43637889

fbshipit-source-id: d2519c28e86e7bed818ba02b7b828d510ba010fe