Identify and fix identifier issues in the IDL reference
Summary:
There is a number of issues with identifiers in the IDL reference:

* A qualified identifier is defined as a syntactic construct (`Name`) while in fact it is (sadly) a lexical one. The distinction is important because it means that whitespace is not allowed between components.
* The name `Name` is too generic and doesn't reflect the fact that it can be (optionally) qualified.
* The distinction between `Name` and `Identifier` is not super clear and there is an obvious inconsistency.
* `Name` is defined in the Namespace Directives section which makes no sense because it is not specific to namespaces.

To fix these issues introduce a lexical symbol, `maybe_qualified_id`, which denotes a qualified or unqualified identifier (in other words, an optionally qualified identifier thus "maybe"). Update `identifier` to denote an unqualified identifier as it was supposed to. More semantically meaningful names for specific uses of `maybe_qualified_id` can be introduced separately as part of cleaning up the grammar for individual constructs. For now just add `namespace_name` to make the namespace directive grammar more readable.

Both Java and C++ have a similar (but syntactic) constructs called *QualifiedIdentifier* and *qualified-id* respectively. In the Java case *QualifiedIdentifier* includes unqualified which seems counter-intuitive so this diff, while adopting the general convention reflects optionality in the naming via "maybe". Possible alternatives are `optionally_qualified_id` or using (`identifier | qualified_id`) none of which seem better.

Reviewed By: iahs

Differential Revision: D43890334

fbshipit-source-id: 1f56c063ada09089b7647f528f0db1851b267c72