hoist_annotated_types codemod
Summary:
In order to migrate unstructured annotations to structured, we have to restrict the positions that can be annotated to those that accept structured annotations, i.e. named nodes. We will allow type annotations only on typedefs and fields, which means annotations on types in all other positions (consts, inside templated types, and in functions) have to be moved. This codemod hoists such annotated types into named typedefs so they can be migrated to structured in a later pass.
Some source locations used in the codemod were missing, so are added in this diff.

Reviewed By: vitaut

Differential Revision: D43957315

fbshipit-source-id: 3c138810bb331b0c7fa6ba4cb1a6f03862399d04