Add typedef structs for conformance
Summary:
This one is a bit more involved than the last testset change because it involved adding a new print method, the old method couldn't really handle typedefs, I kinda hacked this one together by pulling out the code that prints structs and removing a few elements from it.

Also the way that the typedefs are generated doesn't really work like the other unary transforms so I needed to put in some custom logic that overall makes this really hacky, but it works I guess.

Reviewed By: rmakheja

Differential Revision: D43713846

fbshipit-source-id: 3b00bcd2b37214f366b81a2722159cafebc34ce4