adds setting to ping continuously
Summary: Adds a flag to let you run a single test continuously. It will print out the stats ever n seconds using the runtime_s parameter. The stats logic uses this flag so it seemed easy to just keep this in place. This will keep running a test until process is killed.

Reviewed By: AkramaMirza

Differential Revision: D43808068

fbshipit-source-id: cb7f33faa740ea14ddfe9f3f68435d3076e85b51