Fix lints in t_structured::try_append_field
Summary:
`t_structured::try_append_field` takes the field by an rvalue reference but moves it conditionally and the caller often needs to access the field if it wasn't moved. This results in multiple lint errors at each call site, e.g.

```
'field' used after it was moved
fbcode/thrift/compiler/parse/parsing_driver.cc:644:12: move occurred here
      if (!s.try_append_field(std::move(field))) {
             ^
```

This doesn't add any value and can mask real bugs.

Replace the rvalue reference with an lvalue reference which makes more sense anyway since this is effectively an in/out parameter. It is possible to do something more fancy e.g. passing field by value and then returning it on failure but that will make usage more awkward for little benefit as the use-after-move errors in the compiler are easy to debug.

Rename the `elem` argument to `field` while at it.

Reviewed By: iahs

Differential Revision: D44228690

fbshipit-source-id: b31a2f900c2471ab4ea8f7a50095266bcdd9555c