Look for thrift1 first
Summary: On Macbooks it's possible for both thrift and thrift1 to be present, in which case thrift1 is the fbthrift based plugin we want while thrift is Apache thrift. On devservers only thrift is present, so look for thrift1 first to work in both environments. MacBooks will find the right one, and devservers will fallback to the right one.

Reviewed By: avalonalex

Differential Revision: D44298274

fbshipit-source-id: ee03dcbfdad4af2d2a1b0849459f747b61e5bf8e