Fix hex escape sequences
Summary:
Hex escape sequences (`\x...`) are completely broken in Thrift because they are handled in incompatible ways even by supported target languages.

The first problem is that C++, unlike most other languages, consumes arbitrary number of hex digits after `\x`. For example, `\x42cafe` is a single escape sequence in C++, the interpretation of which depends on the compiler. Clang gives an error in this case:

```
.../test_constants.cpp:36:176: error: hex escape sequence ou
t of range
  static folly::Indestructible<::std::vector<::std::string>> const instance{std::initializer_list<::std::string>{apache::thrift:
:StringTraits<std::string>::fromStringLiteral("\x42cafe"),

                                               ^~~~~~~~
```

And, unlike C++, Thrift doesn't have a way to workaround this other than using something different from hex escape sequences which is not always possible.

In Python `\x42cafe` is interpreted as a single escape sequence `\x42` followed by `cafe`.

Hex escape sequences in the Java codegen are handled even worse than in C++. They are translated into `\u00hh` escape sequences which means that they are incorrect for values >= 0x80 which is most of Unicode.

Fix the issue by restricting the number of hex digits after `\x` to exactly two which is already the behavior of Python and Hack:

* https://docs.python.org/3/reference/lexical_analysis.html#string-and-bytes-literals
* https://docs.hhvm.com/hack/source-code-fundamentals/literals

In Java keep the translation logic just for values >= 0xf8 since those are the only once that appear to be used in practice and remove the rest.

This shouldn't be a breaking change in C++ because it will make the code, which previously did not compile, compile in this language. It is a noop in Hack and Python because they require valid UTF-8 which is not affected. Java is the only problematic case but there is only one file where it matters.

Get rid of invalid UTF-8 from our fixtures since it is not something we support in strings.

Additionally diagnose invalid hex escape sequences in the lexer.

Reviewed By: iahs

Differential Revision: D44311140

fbshipit-source-id: 73669bc29abc2d5e4fdc7ba4e299f268e76a3d70