Edit patch.md using inpage editor
Summary:
Static Patch and Dynamic Patch concepts are not limited to C++.

This diff has been automatically generated by the inpage editor.

NOTE: If you want to update this diff, go via the preview link inside the static docs section below.
Ensure you are editing the same page that was used to create this diff.

Reviewed By: iahs

Differential Revision: D44409421

fbshipit-source-id: 64c9f413ac052fc25f8627aa59320e6ab6f7df54