Fix handling of backslash escape sequence
Summary: Complete the move of the `\\` escape sequence logic to the lexer, fixing bugs in Go and JSON generators where they are currently doubled (see fixtures). Simplify Hack and Java generators. In particular, completely remove the horrible `\u00hh` hack from the javadeprecated generator because the only case that relies on this is using javaswift.

Reviewed By: iahs

Differential Revision: D44441791

fbshipit-source-id: 85bad1fc47e766f910b333418721e98ffd7c6526