Tag writes for AsyncSocket::SendMsgParamsCallback
Summary:
For the motivation / mechanism, see the docstring for `struct WriteRequestTag`.

The overall design goals here were:
 - Make it possible to match a `writeChain` call with the corresponding `SendMsgParamsCallback::getAncillaryData` calls (there may be several).
 - Minimize correctness and performance risk to the normal usage of `AsyncSocket`, which is typically free of ancillary data.
 - Avoid introducing new footguns.

For a concrete usage example, see D43935122.

Re -- `AsyncSSLSocket` lacking support: It would be possible to extend `AsyncSSLSocket` to support write tagging, but when I implemented it, we found the extra implementation complexity not to be worth the simpler API contract.

Reviewed By: AkramaMirza

Differential Revision: D43023628

fbshipit-source-id: bd5109a1d32227b67f75794e71d4d288a94b2802