Minor improvements to docs
Summary:
Rename "include statement" to "include directive" because it is a directive to the Thrift compiler and not a statement in the usual sense. Python is an example of a language where `import` **is** a statement, both grammatically and since it represents an action carried out at runtime. We already have the namespace directive which is another compiler directive so it also improves consistency and can potentially simplify some diagnostic messages because we can now talk about directives rather than "include statements and namespace directives".

Remove the manual anchor from "Package Declaration" since we don't do it anywhere else and it confuses some Markdown renderers.

Rename "singular" to "non-container" in the context of Thrift types because the latter seems more intuitive.

Reviewed By: iahs

Differential Revision: D44725633

fbshipit-source-id: 36731e2bd526b4d281843e1bf82880a99506ff24