fix build failure for nested patch
Summary:
Currently if we are using nested patch with `AssignOnlyPatch` annotation, the thrift compiler will crash.

The root cause is that when generating patch for struct and union,

1. We generate the field patch struct first.
2. Then check whehter `AssignOnlyPatch` annotation exists.
3. If `AssignOnlyPatch` annotation doesn't exist, we when set `PatchPrior` and `PatchAfter` fields.

However, for nested patch, since we don't generate `FooPatchPatch`, we will fail to generate field patch struct.

This diff changed the order to check `AssignOnlyPatch` first, then only generate field patch struct if the annotation doesn't exist.

Reviewed By: thedavekwon

Differential Revision: D44694917

fbshipit-source-id: cb7b7b6ac9e0a2855c4b9dd8762259d721b29675