Synchronize CPU-CC Config reads/writes
Summary:
Previously, `CPUConcurrencyController::config()` would evaluate the Config observer on every invocation. This meant that it was possible for a single call to `cycleOnce` to see multiple Config versions. This could put the controller in a weird state until the ObserverManager thread ran the callback to cancel() and re-schedule() the controller loop.

This change wraps the Config in a folly::Synchronized to ensures that:
1. The Config is only updated from the config Observer callback (happens once in the thread that creates the callback and on an ObserverMngr thread after that when observer changes are propagated).
2. The Config doesn't change while executing the controller loop

Reviewed By: joshkirstein

Differential Revision: D44565855

fbshipit-source-id: 2925f4cafb51cf50f91bf32aaf7026abe47d11b9