Construct new context in-place.
Summary:
`emplace` is created to pass constructor arguments so that
new element is created in-place, but in this case an entire Context
rvalue is passed, so instead of in-place construction, move construction
happens. At the moment it's not a big deal because context's move is cheap
but it's a bad practice.

Reviewed By: thedavekwon

Differential Revision: D44812073

fbshipit-source-id: 42087fe76b3d1bcacb1fa008e95d6c8c3eea958d