Change ContextStack associated type from Buffer to Frame
Summary:
Currently `ContextStack`s have an associated `Buffer` type used to represent
serialized requests and responses. This diff changes the associated type to be
`Framing`. With `Framing`, it's harder to log a request as a response (or vice
versa), the included `BufExt` trait bound provides implementations of
`ContextStack` a generic way to read buffers, and the serialized request and
response no longer have to be represented with the same buffer type (which
`Protocol`s do not require).

Because the final framing encoding isn't guaranteed to be buffer-like, e.g. the
sizing serializer's final output is a `usize`, this diff also adds a few type
bounds to require the serialized final output implement `BufExt`. This mirrors
the deserializer's requirement that the serialized request buffer implement
`BufExt`.

Reviewed By: jsgf

Differential Revision: D44441813

fbshipit-source-id: 718f59f3d9a7e304e75846fb376629100b2e0698