Remove BufExt::reset
Summary:
The motivation behind `BufExt::reset` is to be able to read a (de)serialization
buffer more than once, as `Buf` is effectively a read-only iterator that can
only move forward. It's hard to implement for some types, e.g. `Bytes`, though,
as it requires the underlying type record and expose its initial state. `Bytes`
is especially notable because it's the final encoded representation for the
compact, binary, and simple json protocols, and changing the type to
`Cursor<Bytes>` or similar to be able to reset the buffer would change types in
almost all users of fbthrift serialization.

Instead of resetting the buffer, we can `clone` it instead. This gives more
flexibility into what position we can "reset" to, and it's a standard trait
that types should already implement. It's also not unreasonable to require a
buffer implement `Clone` because these are read-only buffers, and so an iterator
over them should be trivially cloneable.

This diff removes `BufExt::reset` and its implementations and replaces `reset`
functionality with `clone`.

Reviewed By: jsgf

Differential Revision:
D44556666

Privacy Context Container: L1122763

fbshipit-source-id: d628b0e41696a2328d4c724ca91e119364e8dc40