Revert to more basic field name generation when py3.hidden is involved
Summary: When using py3.hidden, we can't use `__get_field_name_by_index` since that calls into the cpp object which isn't missing the same fields as the py3 object. Fixing this by just reverting to a more basic field name map. I'm not going to do this more generally because `py3.hidden` is not very widely used and it'd be a fairly large performance regression.

Reviewed By: thedavekwon

Differential Revision: D44772258

fbshipit-source-id: de25d52f63aa175b015f6efa8e9f3e7a6ba3f7a9