In MapPatch::patchByKey(...), if key is removed, ignore the patch to the value
Summary:
If MapPatch removed the key, when user patch this key later, we will still add the patch to the value into MapPatch.
This is unnecessary since if MapPatch removed `key`, the patch to the value will never be applied
This diff changed to return a dummy patch instead so that the patch size is smaller.

Reviewed By: thedavekwon

Differential Revision: D45238997

fbshipit-source-id: 8d377a15407de0e649e2649fe2f2f73c79865e70