Add getter for raw socket
Summary: D35059188 added an observer interface for `RocketServerConnection`. D43335752 allows `NetEngTransportMonitor` to be an observer for `RocketServerConnection` using the ConstructorCallback mechanism. However, to be cautios, we will make TransportMonitor observe the rocket connection only if it already observes the AsyncSocket that drives the connection. This requires us to be have access to the raw socket from outside the connection object. This diff adds a simple getter for the `rawSocket_` member. See D43335752 for a longer discussion on this.

Reviewed By: NiteshKant

Differential Revision: D44318602

fbshipit-source-id: 370fbc1ab39c51f90bfea43dd1b3cfe2016cfe71