Fix boost dependencies in CMake
Summary:
Add a missing `find_package` for Boost and use imported targets instead of CMake variables. Imported targets are a better way of working with dependencies because they correctly propagate various properties such as libraries and include directories.

This fixes compiler-only build which currently fails because include directories are set in a wrong place:

```
.../thrift/compiler/detail/mustache/mstch.h:39:10: fatal error: 'boost/variant.hpp' file not found
```

Reviewed By: avalonalex

Differential Revision: D45369662

fbshipit-source-id: 6f25b956aafa4afd73a8243765c22be939d461ee