Improve parsing of structured annotations
Summary: `parse_structured_annotation` is a bit misleading because it may not parse anything. Change it to always parse an annotation which is more consistent with other methods. It is also more convenient for consumers that don't want to build an AST because they can now return null from `on_structured_annotation`.

Reviewed By: iahs

Differential Revision: D45606590

fbshipit-source-id: 4639961785605e5c1d6ef1e22f5eeb0e57e12676