implement toAny and get methods
Summary:
Currently to convert thrift.Any from/to thrift value, the only API is TypeRegistry.

However, if we have the concrete type, we don't need TypeRegistry since we can serialize/deserialize directly.

In addition, TypeRegistry doesn't handle adapted type, when we convert Patch to Any, we need to use `toThrift()` method.

This diff implemented

1. `toAny<T>(T data)` which converts thrift value to thrift.Any.
2. `get<T>()` which converts thrift.Any to thrift value.

So that thrift.Any can be used without TypeRegistry.

Reviewed By: iahs

Differential Revision: D45523843

fbshipit-source-id: 4cd64aed4d7b8ac7e6a558ff3f2b3b12cbb4d386