Revert D45557696: Multisect successfully blamed D45557696 for test or build failures
Summary:
This diff is reverting D45557696
D45557696: [thrift] Fix for early singleton destruction in MonitoringConnection by srir has been identified to be causing the following test or build failures:

Tests affected:
- [common/tools/thriftdbg/omni_thrift/py/test:client_test - test_exception (common.tools.thriftdbg.omni_thrift.py.test.client_test.TranscoderTest)](https://www.internalfb.com/intern/test/844425010816246/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2017898
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: joshkirstein

Differential Revision: D45683020

fbshipit-source-id: 0364c3baedfed36549a926e3ff7601002f44acb0