Separate types and definitions in the IDL reference
Summary:
Thrift definitions are currently all over the place in the IDL reference and confused with types. For example, enum definitions are described in Basic Thrift Types/Enumeration Types and struct definitions are described in the top-level Struct Types section. Only constant definitions have a reasonably named section but it's separate from everything else. The grammar symbols are also wrong e.g. `StructType` is not a type but a definition.

Fix this by introducing a top-level Definitions section and moving all definitions to it. Fix the grammar symbols making them lowercase and dropping misplaced "Type" where appropriate. This also makes them consistent with the AST names. The content of the sections will be cleaned up in separate diff(s).

Rename the Basic Thrift Types section to Types because it will cover all types.

Reviewed By: iahs

Differential Revision: D45656477

fbshipit-source-id: d97661065337454341e5b23343c2988a88eeb55f