Speed up json protocol string decoding
Summary: By not going through folly's json decoder unless both utf-8 is enabled, and we actually run into a utf escape sequence in the string. This could be narrowed further to only go through Folly if a surrogate pair is found, but I expect utf escape sequences to be rare enough to begin with that it won't matter.

Reviewed By: vitaut

Differential Revision: D44391512

fbshipit-source-id: 44dff11bca4496691171c1236ab6a8ba19c04155