add annotation to support returning anyhow::Error from Rust Thrift service impl
Summary:
handling errors in Rust Thrift server implementations can get rather verbose and is a regular pain point for new Rust Thrift server users.

to help here, open up the option to return `anyhow::Error` inside of Rust Thrift server impl methods. this is kept opt-in as this is generally bad practice, but I think the convenience here outweighs that this is a bad practice?

this mimics what C++ and Python Thrift server do when an unknown exception type is thrown, so this is in line with other Thrift languages.

Reviewed By: chriskonstad

Differential Revision: D45627783

fbshipit-source-id: 061f36726653d3cbf207564198623f2df6097a8e