Fix ThreadManager::newSimpleThreadManager() overload + add priority
Summary:
The ThreadManager::newSimpleThreadManager() overload that supports
named threads did not correctly set up the PosixThreadFactory as the other
overload did (correctly). Fix that bug. At the same time, add support to
specify the priority of the constructed SimpleThreadManager's threads. In some
cases applications may only create a single pool (or a small set of manual
pools) and control over the priority is helpful.

Reviewed By: mikhail-antonov

Differential Revision: D45596594

fbshipit-source-id: 388ec8e409b9acfd126f58140741d369ad38a4ee