Revert D45342650: Multisect successfully blamed D45342650 for test or build failures
Summary:
This diff is reverting D45342650
D45342650: [thrift][server]improve io_uring support by robertroeser has been identified to be causing the following test or build failures:

Tests affected:
- [memcache/memcached:memcached_tests - test_two_pools_shutdown_restart (memcache.memcached.test.test_flat_mc.TestTwoPools)](https://www.internalfb.com/intern/test/562950032989701/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2036550
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: robertroeser

Differential Revision: D45725093

fbshipit-source-id: c111a1f62bec7142717e1089d3d4eb95c6c3dc2b