Support for struct type as map-key
Summary:
When a struct is used as a map key - it needs to be a "concrete" struct, not a pointer:
`map[MyStruct]int64` vs `map[*MyStruct]int64`

However, there is a caveat - a struct may be incomparable (if it contains a slice or a map inside). An incomparable object cannot be used as a map-key. We need to account for that.

If struct is comparable - use "concrete", if struct is not comparable - use pointer (pointers are always comparable).

Using a pointer does not guarantee key uniqueness (because struct contents are not checked for equality) - but the code will be valid and will compile.
Which is a major win over the legacy generator.

Reviewed By: leoleovich

Differential Revision: D45722579

fbshipit-source-id: a7a20203aca4e14d667fa59a56ac18dd1b307e06