Rename `PayloadAndHeader` structs not to be ambiguous
Summary:
Rename two identically-named structs to make it clear that they are different:
 - `RichPayloadReceived` -- formerly `ClientBufferedStream<T>::PayloadAndHeader`, used to receive stream messages
 - `RichPayloadToSend` -- formerly `detail::PayloadAndHeader<T>` aka `ServerStream<T>::PayloadAndHeader`, used to send stream messages

Today, these are not the same type, which works out because:
 - the "send" variety needs to be copiable for "multi-publisher" patterns -- when I add FD-passing support, this will store `vector<shared_ptr<File>>`
 - the "receive" variety only need to be movable -- and when I add FD-passing support, it must support sole `File` ownership (i.e. `vector<File>`).

This parallels the split of `SocketFds::{ToSend,Received}`.

Additionally, once I add FD support, these will no longer be a "payload, header" pair, so the new name reflects that.

Reviewed By: AkramaMirza

Differential Revision: D45675818

fbshipit-source-id: 9b6076a14b7c8bfd5a64c17aeedac266f94b92b7