Cleanup parser
Summary:
Rename "annotation" to "deprecated_annotation" in more places to avoid ambiguity with new (structured) annotations. This also makes it easier to find all places where we still use deprecated annotations.

Mark the identifier corresponding to the function name as non-optional in the function declaration grammar. The implementation already treats it as such.

Reviewed By: Mizuchi

Differential Revision: D45877988

fbshipit-source-id: dd1bd1701c091fd7bc7b2a8bee1b7331ad6e16c0