Remove a problematic annotation
Summary:
The `thrift.GenDefaultEnumValue` annotation adds a bunch of complexity for little value. We already have a language mechanism for this which is recognized by anyone who has experience with any programming language that has enums.

Instead, `thrift.GenDefaultEnumValue` requires understanding annotations, the semantics of this specific annotation and adds a dependency on the annotation library just to define a hidden enumerator. What's even worse, it also introduces some codegen magic and another mechanism to define enumerators.

Remove the few remaining usages of this questionable annotation and replace the annotation itself with a stub to be removed while all syncs are complete. This makes the intent much clearer and follows the usual recommendation from the IDL reference: https://www.internalfb.com/intern/staticdocs/thrift/docs/idl/#enumerations.

As a nice side-effect this removes ~4 kLOC of noise from fixtures. Additional 11 kLOC will be removed when the annotations is completely gone.

Reviewed By: iahs

Differential Revision: D45964721

fbshipit-source-id: b5d0c5f4624036df2ce1046186c3eb72ab9bf128